Initial commit
Basic example of the problem.

With the toolchain configured in rust-toolchain.toml, it works fine, and
the device prints values in the hundreds.
With the other toolchain (the one commented out), it reads mostly 0,
with some other very small values occasionally.

The first behavior is correct, suggesting some kind of compiler
regression.